*** Important: Use better scaled initialization!
This initialization scale comes from the coursera deep learning course. Take layer 1 for example, the rational is to select a scale factor based on the number of nodes in the previous layer (input layer). So that the summation (Z1) is properly scaled  and close to (-1, 1). Too small or two large won't be good.